Fix unread notification count due to promises
